requirements.txt
pystache>=0.5.4
protobuf>=3.3
google-gax>=0.12.3
pyyaml>=3.12
